working on paging, got the basics for the pmm

+++ src/lib/core/core.h
@@ -42,7 +42,7 @@
 typedef unsigned char       uint8_t;
 typedef unsigned short      uint16_t;
 typedef unsigned int        uint32_t;
-typedef uint16_t            size_t;
+typedef unsigned int        size_t;
 
 
 typedef unsigned char       uint8;
